Change int to long in SalesManager

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[]arr = {1, 20, 50, 70};
+        long[]arr = {1, 20, 50, 70};
         SalesManager salesManager = new SalesManager(arr);
-        int max = salesManager.max();
+        long max = salesManager.max();
 
         System.out.println(max);
     }
